fix: 라이브러리로써 올바르게 import될 수 있도록 수정
